Update New-MailboxImportRequest.md
We might want to go so far as to remove Exchange Online from the list of supported products.  I cannot find a way to successfully use this cmdlet with EXO.  The EXO implementation lacks the Filepath argument and Exchange Product Group does not support the use of the cmdlet with the SAS key provided by the PST Ingestion Service.  The cmdlet should probably be removed from the RBAC configuration as well.